Split `AssetsTest` class
